@@ -5,27 +5,24 @@ declare(strict_types=1);
 namespace AutoMapper\Tests;
 
 use AutoMapper\AutoMapper;
+use AutoMapper\Extractor\ClassMethodToCallbackExtractor;
 use AutoMapper\Generator\Generator;
 use AutoMapper\Loader\ClassLoaderInterface;
 use AutoMapper\Loader\FileLoader;
-use Doctrine\Common\Annotations\AnnotationReader;
 use PhpParser\ParserFactory;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Serializer\Mapping\ClassDiscriminatorFromClassMetadata;
 use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory;
-use Symfony\Component\Serializer\Mapping\Loader\AnnotationLoader;
+use Symfony\Component\Serializer\Mapping\Loader\AttributeLoader;
 
 /**
  * @author Baptiste Leduc <baptiste.leduc@gmail.com>
  */
 abstract class AutoMapperBaseTest extends TestCase
 {
-    /** @var AutoMapper */
-    protected $autoMapper;
-
-    /** @var ClassLoaderInterface */
-    protected $loader;
+    protected AutoMapper $autoMapper;
+    protected ClassLoaderInterface $loader;
 
     protected function setUp(): void
     {
@@ -37,9 +34,10 @@ abstract class AutoMapperBaseTest extends TestCase
     {
         $fs = new Filesystem();
         $fs->remove(__DIR__ . '/cache/');
-        $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader(new AnnotationReader()));
+        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
 
         $this->loader = new FileLoader(new Generator(
+            new ClassMethodToCallbackExtractor(),
             (new ParserFactory())->create(ParserFactory::PREFER_PHP7),
             new ClassDiscriminatorFromClassMetadata($classMetadataFactory),
             $allowReadOnlyTargetToPopulate
