@@ -13,31 +13,31 @@ use AutoMapper\MapperGeneratorMetadataFactoryInterface;
 use AutoMapper\Transformer\ArrayTransformerFactory;
 use AutoMapper\Transformer\BuiltinTransformerFactory;
 use AutoMapper\Transformer\ChainTransformerFactory;
+use AutoMapper\Transformer\CustomTransformer\CustomTransformersRegistry;
 use AutoMapper\Transformer\DateTimeTransformerFactory;
 use AutoMapper\Transformer\MultipleTransformerFactory;
 use AutoMapper\Transformer\NullableTransformerFactory;
 use AutoMapper\Transformer\ObjectTransformerFactory;
 use AutoMapper\Transformer\UniqueTypeTransformerFactory;
-use Doctrine\Common\Annotations\AnnotationReader;
 use Symfony\Component\PropertyInfo\Extractor\PhpDocExtractor;
 use Symfony\Component\PropertyInfo\Extractor\ReflectionExtractor;
 use Symfony\Component\PropertyInfo\PropertyInfoExtractor;
 use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory;
-use Symfony\Component\Serializer\Mapping\Loader\AnnotationLoader;
+use Symfony\Component\Serializer\Mapping\Loader\AttributeLoader;
 
 /**
  * @author Baptiste Leduc <baptiste.leduc@gmail.com>
  */
 class MapperGeneratorMetadataFactoryTest extends AutoMapperBaseTest
 {
-    /** @var MapperGeneratorMetadataFactoryInterface */
-    protected $factory;
+    protected MapperGeneratorMetadataFactoryInterface $factory;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader(new AnnotationReader()));
+        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
+        $customTransformerRegistry = new CustomTransformersRegistry();
         $reflectionExtractor = new ReflectionExtractor(null, null, null, true, ReflectionExtractor::ALLOW_PUBLIC | ReflectionExtractor::ALLOW_PROTECTED | ReflectionExtractor::ALLOW_PRIVATE);
 
         $phpDocExtractor = new PhpDocExtractor();
@@ -54,6 +54,7 @@ class MapperGeneratorMetadataFactoryTest extends AutoMapperBaseTest
             $reflectionExtractor,
             $reflectionExtractor,
             $transformerFactory,
+            $customTransformerRegistry,
             $classMetadataFactory
         );
 
@@ -62,6 +63,7 @@ class MapperGeneratorMetadataFactoryTest extends AutoMapperBaseTest
             $reflectionExtractor,
             $reflectionExtractor,
             $transformerFactory,
+            $customTransformerRegistry,
             $classMetadataFactory
         );
 
@@ -70,6 +72,7 @@ class MapperGeneratorMetadataFactoryTest extends AutoMapperBaseTest
             $reflectionExtractor,
             $reflectionExtractor,
             $transformerFactory,
+            $customTransformerRegistry,
             $classMetadataFactory
         );
 
