@@ -14,11 +14,11 @@ interface AutoMapperInterface
     /**
      * Maps data from a source to a target.
      *
-     * @param array|object|null   $source  Any data object, which may be an object or an array
-     * @param string|array|object $target  To which type of data, or data, the source should be mapped
-     * @param array               $context Mapper context
+     * @param array|object                      $source  Any data object, which may be an object or an array
+     * @param 'array'|class-string|array|object $target  To which type of data, or data, the source should be mapped
+     * @param array                             $context Mapper context
      *
      * @return array|object|null The mapped object
      */
-    public function map(null|array|object $source, string|array|object $target, array $context = []): null|array|object;
+    public function map(array|object $source, string|array|object $target, array $context = []): null|array|object;
 }
