@@ -6,6 +6,7 @@ namespace AutoMapper\Extractor;
 
 use AutoMapper\Exception\InvalidMappingException;
 use AutoMapper\MapperMetadataInterface;
+use AutoMapper\Transformer\CustomTransformer\CustomTransformersRegistry;
 use AutoMapper\Transformer\TransformerFactoryInterface;
 use Symfony\Component\PropertyInfo\PropertyInfoExtractorInterface;
 use Symfony\Component\PropertyInfo\PropertyReadInfo;
@@ -32,10 +33,11 @@ final class FromTargetMappingExtractor extends MappingExtractor
         PropertyReadInfoExtractorInterface $readInfoExtractor,
         PropertyWriteInfoExtractorInterface $writeInfoExtractor,
         TransformerFactoryInterface $transformerFactory,
+        CustomTransformersRegistry $customTransformerRegistry,
         ClassMetadataFactoryInterface $classMetadataFactory = null,
-        private ?AdvancedNameConverterInterface $nameConverter = null,
+        private readonly ?AdvancedNameConverterInterface $nameConverter = null,
     ) {
-        parent::__construct($propertyInfoExtractor, $readInfoExtractor, $writeInfoExtractor, $transformerFactory, $classMetadataFactory);
+        parent::__construct($propertyInfoExtractor, $readInfoExtractor, $writeInfoExtractor, $transformerFactory, $customTransformerRegistry, $classMetadataFactory);
     }
 
     public function getPropertiesMapping(MapperMetadataInterface $mapperMetadata): array
@@ -61,10 +63,15 @@ final class FromTargetMappingExtractor extends MappingExtractor
             $sourceTypes = [];
 
             foreach ($targetTypes as $type) {
-                $sourceTypes[] = $this->transformType($mapperMetadata->getSource(), $type);
+                $sourceType = $this->transformType($mapperMetadata->getSource(), $type);
+
+                if ($sourceType) {
+                    $sourceTypes[] = $sourceType;
+                }
             }
 
-            $transformer = $this->transformerFactory->getTransformer($sourceTypes, $targetTypes, $mapperMetadata);
+            $transformer = $this->customTransformerRegistry->getCustomTransformerClass($mapperMetadata, $sourceTypes, $targetTypes, $property)
+                ?? $this->transformerFactory->getTransformer($sourceTypes, $targetTypes, $mapperMetadata);
 
             if (null === $transformer) {
                 continue;
@@ -86,7 +93,7 @@ final class FromTargetMappingExtractor extends MappingExtractor
                 $this->getMaxDepth($mapperMetadata->getTarget(), $property),
                 $this->isIgnoredProperty($mapperMetadata->getSource(), $property),
                 $this->isIgnoredProperty($mapperMetadata->getTarget(), $property),
-                PropertyReadInfo::VISIBILITY_PUBLIC === $this->readInfoExtractor->getReadInfo($mapperMetadata->getSource(), $property)?->getVisibility() ?? true,
+                PropertyReadInfo::VISIBILITY_PUBLIC === ($this->readInfoExtractor->getReadInfo($mapperMetadata->getSource(), $property)?->getVisibility() ?? PropertyReadInfo::VISIBILITY_PUBLIC),
             );
         }
 
