@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace AutoMapper\Extractor;
 
+use AutoMapper\Transformer\CustomTransformer\CustomTransformersRegistry;
 use AutoMapper\Transformer\TransformerFactoryInterface;
 use Symfony\Component\PropertyInfo\PropertyInfoExtractorInterface;
 use Symfony\Component\PropertyInfo\PropertyReadInfo;
@@ -25,6 +26,7 @@ abstract class MappingExtractor implements MappingExtractorInterface
         protected readonly PropertyReadInfoExtractorInterface $readInfoExtractor,
         protected readonly PropertyWriteInfoExtractorInterface $writeInfoExtractor,
         protected readonly TransformerFactoryInterface $transformerFactory,
+        protected readonly CustomTransformersRegistry $customTransformerRegistry,
         private readonly ?ClassMetadataFactoryInterface $classMetadataFactory = null,
     ) {
     }
@@ -112,10 +114,6 @@ abstract class MappingExtractor implements MappingExtractorInterface
             return null;
         }
 
-        if (!$this->classMetadataFactory->getMetadataFor($class)) {
-            return null;
-        }
-
         $serializerClassMetadata = $this->classMetadataFactory->getMetadataFor($class);
         $maxDepth = null;
 
@@ -134,7 +132,7 @@ abstract class MappingExtractor implements MappingExtractorInterface
             return null;
         }
 
-        if (null === $this->classMetadataFactory || !$this->classMetadataFactory->getMetadataFor($class)) {
+        if (null === $this->classMetadataFactory) {
             return null;
         }
 
@@ -167,7 +165,7 @@ abstract class MappingExtractor implements MappingExtractorInterface
             return false;
         }
 
-        if (null === $this->classMetadataFactory || !$this->classMetadataFactory->getMetadataFor($class)) {
+        if (null === $this->classMetadataFactory) {
             return false;
         }
 
