@@ -149,6 +149,8 @@ class MapperMetadata implements MapperGeneratorMetadataInterface
             $hash .= filemtime($reflection->getFileName());
         }
 
+        $hash .= AutoMapper::VERSION_ID;
+
         return $hash;
     }
 
@@ -229,7 +231,8 @@ class MapperMetadata implements MapperGeneratorMetadataInterface
                 null,
                 new CallbackTransformer($property),
                 $property,
-                false
+                false,
+                isPublic: true,
             );
         }
     }
