@@ -17,7 +17,7 @@ use Symfony\Component\PropertyInfo\Type;
  *
  * @author Joel Wurtz <jwurtz@jolicode.com>
  */
-final class BuiltinTransformer implements TransformerInterface
+final readonly class BuiltinTransformer implements TransformerInterface
 {
     private const CAST_MAPPING = [
         Type::BUILTIN_TYPE_BOOL => [
@@ -69,22 +69,32 @@ final class BuiltinTransformer implements TransformerInterface
             return $type->getBuiltinType();
         }, $this->targetTypes);
 
-        // Source type is in target => no cast
         if (\in_array($this->sourceType->getBuiltinType(), $targetTypes, true)) {
+            /* Output type is the same as input type so we simply return the same expression */
             return [$input, []];
         }
 
-        // Cast needed
         foreach (self::CAST_MAPPING[$this->sourceType->getBuiltinType()] as $castType => $castMethod) {
             if (\in_array($castType, $targetTypes, true)) {
                 if (method_exists($this, $castMethod)) {
+                    /*
+                     * Use specific cast expression if callback exist in this class
+                     *
+                     * $array = [$source->property];
+                     */
                     return [$this->$castMethod($input), []];
                 }
 
+                /*
+                 * Use the cast expression find in the cast matrix
+                 *
+                 * $bool = (bool) $source->int;
+                 */
                 return [new $castMethod($input), []];
             }
         }
 
+        /* When there is no possibility to cast we assume that the mutator will be able to handle the value */
         return [$input, []];
     }
 
