@@ -18,7 +18,7 @@ use Symfony\Component\PropertyInfo\Type;
  *
  * @author Joel Wurtz <jwurtz@jolicode.com>
  */
-final class ObjectTransformer implements TransformerInterface, DependentTransformerInterface, AssignedByReferenceTransformerInterface
+final readonly class ObjectTransformer implements TransformerInterface, DependentTransformerInterface, AssignedByReferenceTransformerInterface
 {
     public function __construct(
         private Type $sourceType,
@@ -30,6 +30,11 @@ final class ObjectTransformer implements TransformerInterface, DependentTransfor
     {
         $mapperName = $this->getDependencyName();
 
+        /*
+         * Use a sub mapper to map the property
+         *
+         * $this->mappers['Mapper_SourceType_TargetType']->map($input, MapperContext::withNewContext($context, $propertyMapping->property));
+         */
         return [new Expr\MethodCall(new Expr\ArrayDimFetch(
             new Expr\PropertyFetch(new Expr\Variable('this'), 'mappers'),
             new Scalar\String_($mapperName)
@@ -62,6 +67,11 @@ final class ObjectTransformer implements TransformerInterface, DependentTransfor
         $sourceTypeName = 'array';
 
         if (Type::BUILTIN_TYPE_OBJECT === $this->sourceType->getBuiltinType()) {
+            /**
+             * Cannot be null since we check the source type is an Object.
+             *
+             * @var string $sourceTypeName
+             */
             $sourceTypeName = $this->sourceType->getClassName();
         }
 
@@ -73,6 +83,11 @@ final class ObjectTransformer implements TransformerInterface, DependentTransfor
         $targetTypeName = 'array';
 
         if (Type::BUILTIN_TYPE_OBJECT === $this->targetType->getBuiltinType()) {
+            /**
+             * Cannot be null since we check the target type is an Object.
+             *
+             * @var string $targetTypeName
+             */
             $targetTypeName = $this->targetType->getClassName();
         }
 
