@@ -10,9 +10,8 @@ class Foo
 {
     /**
      * @var int
-     *
-     * @Groups({"group1", "group2", "group3"})
      */
+    #[Groups(['group1', 'group2', 'group3'])]
     private $id = 0;
 
     public function getId(): int
